% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_agq.R
\name{bbmix_lkl_agq_fixvar}
\alias{bbmix_lkl_agq_fixvar}
\title{Log-Likelihood after removing the choose(n,y) terms (1/sqrt(2*pi) terms are effectively preserved)}
\usage{
bbmix_lkl_agq_fixvar(
  b,
  sigma2,
  phi,
  y,
  n,
  X,
  subjfac,
  niter_laplace,
  num.nodes,
  sumlkl = TRUE
)
}
\arguments{
\item{If}{num.nodes=1, likelihood is approximated by Laplace method. num.nodes>1 is deprecated.}
}
\description{
Log-Likelihood after removing the choose(n,y) terms (1/sqrt(2*pi) terms are effectively preserved)
}
