% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmix_vem}
\alias{bbmix_vem}
\title{Fit beta-binomial random-effects model using variational EM algorithm}
\usage{
bbmix_vem(
  param,
  y,
  n,
  X,
  subj,
  niter,
  niter_laplace,
  num.nodes,
  optim.method = "BFGS",
  converge_tol = 1e-06
)
}
\description{
Fit beta-binomial random-effects model using variational EM algorithm
}
