% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmix_vem_estep}
\alias{bbmix_vem_estep}
\title{Beta-binomial model VEM E-step}
\usage{
bbmix_vem_estep(
  linpred,
  sigma2,
  phi,
  y,
  n,
  subjfac,
  randint,
  randint.prec,
  ghq,
  niter_laplace
)
}
\description{
randint, randint.prec are initial values
}
