% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem_fixs2.R
\name{bbmix_vem_mstep_fixs2}
\alias{bbmix_vem_mstep_fixs2}
\title{Beta-binomial model VEM M-step}
\usage{
bbmix_vem_mstep_fixs2(
  param,
  sigma2,
  y,
  n,
  X,
  subjfac,
  randint,
  randint.prec,
  ghq,
  optim.method
)
}
\arguments{
\item{param:}{b and phi}
}
\description{
Beta-binomial model VEM M-step
}
