% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmix_vem_qfunc}
\alias{bbmix_vem_qfunc}
\title{VEM Q function - only terms which are functions of b and phi are preserved}
\usage{
bbmix_vem_qfunc(param, sigma2, y, n, X, subjfac, randint, randint.prec, ghq)
}
\description{
VEM Q function - only terms which are functions of b and phi are preserved
}
