% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmixture_lkl_agq}
\alias{bbmixture_lkl_agq}
\title{Likelihood function of mixture model}
\usage{
bbmixture_lkl_agq(b, sigma2, phi, p, y, n, X, subj, niter_laplace, num.nodes)
}
\description{
Likelihood function of mixture model
}
