% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmixture_vem}
\alias{bbmixture_vem}
\title{Mixture model VEM wrapper function}
\usage{
bbmixture_vem(
  param,
  y,
  n,
  X,
  subj,
  niter,
  niter_laplace,
  num.nodes,
  optim.method = "BFGS",
  converge_tol = 1e-06
)
}
\description{
param is initial parameters
}
