% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmixture_vem.R
\name{bbmixture_vem_estep}
\alias{bbmixture_vem_estep}
\title{Mixture model VEM E-step}
\usage{
bbmixture_vem_estep(
  linpred,
  sigma2,
  phi,
  p,
  y,
  n,
  subjfac,
  randint,
  randint.prec,
  ghq,
  niter_laplace
)
}
\description{
randint, randint.prec are initial values
}
