% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_vem.R
\name{bbmixture_vem_mstep}
\alias{bbmixture_vem_mstep}
\title{Mixture model VEM M-step}
\usage{
bbmixture_vem_mstep(
  param,
  wt,
  y,
  n,
  X,
  subjfac,
  randint,
  randint.prec,
  ghq,
  optim.method
)
}
\description{
Mixture model VEM M-step
}
