% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix_agq.R
\name{compute_randint_deriv}
\alias{compute_randint_deriv}
\title{Compute first and second derivatives of
sum(lbeta(mu/phi+y,(1-mu)/phi+n-y)-lbeta(mu/phi,(1-mu)/phi)) wrt random intercept
length(randint)==nsubj, the k-th element corresponds to subjfac==k}
\usage{
compute_randint_deriv(
  linpred,
  randint,
  subjfac,
  y,
  n,
  phi,
  include_deriv0 = FALSE
)
}
\description{
Compute first and second derivatives of
sum(lbeta(mu/phi+y,(1-mu)/phi+n-y)-lbeta(mu/phi,(1-mu)/phi)) wrt random intercept
length(randint)==nsubj, the k-th element corresponds to subjfac==k
}
