% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmixture_vem.R
\name{logBB_bysubj}
\alias{logBB_bysubj}
\title{Mixture model VEM mixture weights}
\usage{
logBB_bysubj(linpred, sigma2, phi, y, n, subjfac, randint, randint.prec, ghq)
}
\description{
Mixture model VEM mixture weights
}
